% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenRange_get_license.R
\name{OpenRange_get_license}
\alias{OpenRange_get_license}
\title{Download license information}
\usage{
OpenRange_get_license(...)
}
\arguments{
\item{...}{Additional arguments passed to internal functions.}
}
\value{
Data.frame containing information on the license associated with the maps.
}
\description{
OpenRange_get_license provides information on model licenses.
}
\examples{
\dontrun{

range_license_info <- OpenRange_get_license()

}
}
\seealso{
Other range functions: 
\code{\link{OpenRange_api_species}()},
\code{\link{OpenRange_get_stats}()},
\code{\link{OpenRange_list_scenarios}()},
\code{\link{OpenRange_load_species}()},
\code{\link{OpenRange_sf}()},
\code{\link{OpenRange_species}()}

Other metadata functions: 
\code{\link{OpenRange_get_stats}()},
\code{\link{OpenRange_list_scenarios}()}
}
\concept{metadata functions}
\concept{range functions}
