% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenRange_get_stats.R
\name{OpenRange_get_stats}
\alias{OpenRange_get_stats}
\title{Download model performance statistics}
\usage{
OpenRange_get_stats(...)
}
\arguments{
\item{...}{Additional arguments passed to internal functions.}
}
\value{
List object containing one data.frame for each of the three modelling frameworks.
}
\description{
OpenRange_get_stats provides information on model performance.
}
\examples{
\donttest{

range_stats <- OpenRange_get_stats()

}
}
\seealso{
Other range functions: 
\code{\link{OpenRange_api_load_species}()},
\code{\link{OpenRange_api_species}()},
\code{\link{OpenRange_get_license}()},
\code{\link{OpenRange_list_scenarios}()},
\code{\link{OpenRange_load_species}()},
\code{\link{OpenRange_sf}()},
\code{\link{OpenRange_species}()}

Other metadata functions: 
\code{\link{OpenRange_get_license}()},
\code{\link{OpenRange_list_scenarios}()}
}
\concept{metadata functions}
\concept{range functions}
