% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenRange.R
\name{OpenRange_list_scenarios}
\alias{OpenRange_list_scenarios}
\title{Download information on available climate change scenarios}
\usage{
OpenRange_list_scenarios(...)
}
\arguments{
\item{...}{Additional arguments passed to internal functions.}
}
\value{
Data.frame containing climate change scenarios available in BIEN.
}
\description{
OpenRange_list_scenarios provides information on climate scenarios with range projections available.
}
\note{
Details on the construction of BIEN range maps is available at http://bien.nceas.ucsb.edu/bien/biendata/bien-3/
}
\examples{
\dontrun{

cc_scenarios <- OpenRange_list_scenarios()

}
}
\seealso{
Other range functions: 
\code{\link{OpenRange_api_species}()},
\code{\link{OpenRange_get_license}()},
\code{\link{OpenRange_get_stats}()},
\code{\link{OpenRange_load_species}()},
\code{\link{OpenRange_sf}()},
\code{\link{OpenRange_species}()}

Other metadata functions: 
\code{\link{OpenRange_get_license}()},
\code{\link{OpenRange_get_stats}()}
}
\concept{metadata functions}
\concept{range functions}
