% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges_sql.R
\name{ranges_sql}
\alias{ranges_sql}
\title{Run an SQL query on the OpenRange database.}
\usage{
ranges_sql(
  query,
  user = "public_bien",
  password = "bien_public",
  ranges = NULL,
  limit = NULL,
  return.query = FALSE,
  schema = "analytical_db",
  db_name = "vegbien",
  print.query = FALSE
)
}
\arguments{
\item{query}{A PostgreSQL query.}

\item{user}{The username used to access the BIEN database}

\item{password}{The password associated with the username}

\item{ranges}{Alternative value to be substituted for "ranges" in queries when not NULL.}

\item{limit}{A limit on the number of records to be returned.  Should be a single number or NULL (the default).}

\item{return.query}{Should  the query used be returned rather than executed?  Default is FALSE}

\item{schema}{An alternative schema to be accessed.  Used for testing purposes.}

\item{db_name}{An alternate database to be used rather than the default, vegbien.}

\item{print.query}{Should  the query used be printed?  Default is FALSE}
}
\value{
A dataframe returned by the query.
}
\description{
ranges_sql is an internal function used to submit SQL queries.
}
\examples{
\dontrun{
ranges_sql("SELECT DISTINCT  scenario_id, scenario,
scenario_filecode, time_period, climate_model, rcp
FROM ranges.range ;")}
}
\keyword{internal}
