% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenRange_api.R
\name{OpenRange_api_species}
\alias{OpenRange_api_species}
\title{Download range maps for given species.}
\usage{
OpenRange_api_species(species, directory = NULL, include_id = TRUE, ...)
}
\arguments{
\item{species}{A single species.}

\item{directory}{Directory that range maps should be saved in.  If none is specified, range maps will be saved in the current working directory.}

\item{include_id}{Logical. Should the range_id be appended to the file name?  Needed to save multiple maps per species.}

\item{...}{Additional arguments passed to internal functions.}
}
\value{
Range maps for specified species.
}
\description{
OpenRange_species extracts range maps for the specified species.
}
\note{
Details on the construction of BIEN range maps is available at http://bien.nceas.ucsb.edu/bien/biendata/bien-3/
}
\examples{
\dontrun{

library(maps) #a convenient source of maps

temp_dir <- file.path(tempdir(), "BIEN_temp")

#Download ranges
OpenRange_api_species(species = "Abies_amabilis",
                 directory = temp_dir)#saves ranges to a temporary directory

#Reading files

Acer_poly <- st_read(dsn = temp_dir,
                     layer = "Abies_amabilis_117684")

#Plotting files
plot(Abies_poly[,1])#plots the range, but doesn't mean much without any reference
map('world', fill = TRUE, col = "grey")#plots a world map (WGS84 projection), in grey
plot(Abies_poly,col="forest green",add=TRUE) #adds the range of Abies lasiocarpa to the map

#Getting data from the files
Abies_poly \%>\%
 st_drop_geometry()

}
}
\seealso{
Other range functions: 
\code{\link{OpenRange_get_license}()},
\code{\link{OpenRange_get_stats}()},
\code{\link{OpenRange_list_scenarios}()},
\code{\link{OpenRange_load_species}()},
\code{\link{OpenRange_sf}()},
\code{\link{OpenRange_species}()}
}
\concept{range functions}
